requirements.txt
streamlit
openai
google-generativeai
python-dotenv
requests
pillow
transformers
torch